pythonapp/pyproject.toml
[project]
name = "kctest"
version = "0.1.0"
description = "Keycloak Test Client"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-oidc>=2.4.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
]
